+++ src/color.cpp
@@ -3,8 +3,8 @@
 #include "vector.h"
 #include "color.h"
 
-CUDA_CALLABLE Color::Color(double r, double g, double b)
-	: Vector3d<double>(r, g, b)
+CUDA_CALLABLE Color::Color(float r, float g, float b)
+	: Vector3d<float>(r, g, b)
 {
 }
 
@@ -12,7 +12,7 @@ CUDA_CALLABLE Color::~Color()
 {
 }
 
-CUDA_CALLABLE double* Color::RGB() const
+CUDA_CALLABLE float* Color::RGB() const
 {
 	return this->values();
 }
