+++ src/scene.cpp
@@ -4,19 +4,19 @@
 
 Scene::Scene(Point camera_location) 
 	:	camera_location_(camera_location), 
-		image_bl_location_(0.,0.,0.)
+		image_bl_location_(0.f,0.f,0.f)
 {
-	double ar = 16./9.; // aspect ration = width/height
+	float ar = 16.f/9.f; // aspect ration = width/height
 	int height = 1000;
 	int width = (int)(height*ar);
 	image_ = new Image(height, width);
 
 	// we align center of image with camera
 	// image is at z=1 relative to camera
-	image_dims_[0] = 2.;
+	image_dims_[0] = 2.f;
 	image_dims_[1] = image_dims_[0]/ar;
-	double image_z = +1;
-	image_bl_location_ = camera_location_ + Point(-image_dims_[0]/2., -image_dims_[1]/2., image_z);
+	float image_z = +1.f;
+	image_bl_location_ = camera_location_ + Point(-image_dims_[0]/2.f, -image_dims_[1]/2.f, image_z);
 }
 
 Scene::~Scene()
@@ -48,7 +48,7 @@ CUDA_CALLABLE Point Scene::pixel_location(int irow, int jcol) const
 	// scale it by the image's actual dimensions
 	// before computng its absolute location
 	Point p = image()->pixel_location(irow, jcol);
-	double* vals = p.values();
+	float* vals = p.values();
 	vals[0] *= image_dims_[0];
 	vals[1] *= image_dims_[1];
 	//printf("%d %d: %f %f %f \n", irow, jcol, (image_bl_location_ + p).values()[0], (image_bl_location_ + p).values()[1], (image_bl_location_ + p).values()[2]);
@@ -75,15 +75,15 @@ CUDA_CALLABLE Light* Scene::light(int i) const
 	return lights_[i];
 }
 
-CUDA_CALLABLE Sphere* Scene::object_hit(const Ray& ray,  Point& hit_location, Vector3d<double>& normal_hit_dir) const
+CUDA_CALLABLE Sphere* Scene::object_hit(const Ray& ray,  Point& hit_location, Vector3d<float>& normal_hit_dir) const
 {
 	Sphere* closest_obj_hit = nullptr;
-	double min_dist = INFINITY;
+	float min_dist = INFINITY;
 
 	const int n = objects_.size();
 	for (int i = 0; i < n; i++) {
 		Point location;
-		double dist = objects_[i]->hit_distance(*object_locations_[i], ray, location);
+		float dist = objects_[i]->hit_distance(*object_locations_[i], ray, location);
 		if (dist < min_dist) {
 			min_dist = dist;
 			closest_obj_hit = objects_[i];
